헤더 및 드로어 컴포넌트 리팩토링
